requirements.txt
# Core dependencies
PyQt5>=5.15.0
numpy>=1.19.0
pyinstaller>=5.0.0

# Video processing
moviepy>=2.0.0              # For video export/cutting

# Video player
python-vlc>=3.0.0           # VLC player for video preview with audio

# CSV/Excel import functionality
pandas>=1.3.0
openpyxl>=3.0.0
